requirements.txt
Flask
numpy
tensorflow
pymongo
keras
dnspython
gunicorn
gdown
python-dotenv